Moved TouchablePlatformSpecific to main component export
